Working with new server
Trying to use the DDD
