+++ Bubble Messenger/Scripts/Server/Domain/IIdProvider.cs
@@ -1,0 +1,7 @@
+namespace Domain
+{
+    public interface IIdProvider
+    {
+        public string GetId(int length);
+    }
+}
