+++ Bubble Messenger/Scripts/Server/Domain/Message_repo.cs
@@ -1,0 +1,13 @@
+namespace Domain
+{
+    interface IMessageRepo
+    {
+        public void GetMessage(string id);
+
+        public void CreateMessage(User sender, Dialogue receiver, string text, string time, string timestamp, string json);
+
+        public void DeleteMessage(string id);
+
+        public void UpdateMessage(Message message);
+    }
+}
