+++ Bubble Messenger/Scripts/Server/Domain/User_repo.cs
@@ -1,0 +1,13 @@
+namespace Domain
+{
+    interface IUserRepo
+    {
+        public User GetUser(string id);
+
+        public void CreateUser(string name, string password, string bio);
+
+        public void DeleteUser(string id);
+
+        public void UpdateUser(User user);
+    }
+}
