+++ Bubble Messenger/Scripts/Server/Infrastructure/RandomIdProvider.cs
@@ -1,0 +1,21 @@
+﻿using Domain;
+
+namespace Infrastructure
+{
+    public class RandomIdProvider : IIdProvider
+    {
+        public string GetId(int length)
+        {
+            const string chars = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789";
+            Random random = new Random();
+            char[] result = new char[length];
+
+            for (int i = 0; i < length; i++)
+            {
+                result[i] = chars[random.Next(chars.Length)];
+            }
+
+            return new string(result);
+        }
+    }
+}
